% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculateArrivalAndDepartureTime}
\alias{calculateArrivalAndDepartureTime}
\title{Calculate station arrival and departure time}
\usage{
calculateArrivalAndDepartureTime(
  startTime,
  transitTime,
  operationTime,
  extraOperationTime
)
}
\arguments{
\item{startTime}{a \code{POSIXct} value indicating the start time of the cruise.}

\item{transitTime}{a numeric vector of calculated transit times in hours.}

\item{operationTime}{a numeric vector of total operation time in hours.}

\item{extraOperationTime}{a numeric vector of extra operation time in hours.}
}
\description{
Calculate station arrival and departure time
}
\author{
Chantelle Layton
}
