% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculateBottomDepth}
\alias{calculateBottomDepth}
\title{Infer bottom depth at stations.}
\usage{
calculateBottomDepth(longitude, latitude)
}
\arguments{
\item{longitude}{A numeric value indicating longitude in decimal degrees.}

\item{latitude}{A numeric value indicating latitude in decimal degrees.}
}
\description{
Infer bottom depth at stations by downloading ETOPO1 dataset using the \code{oce::download.topo()} function.
The topographic data is saved as a netCDF file in the local working directory, and on subsequent calls to
\code{oce::download.topo()} with identical parameters, it will simply return the cached file. See the help page for \code{oce:download.topo()}
for additional details.
}
\author{
Chantelle Layton
}
