% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculateDistanceToNextStation}
\alias{calculateDistanceToNextStation}
\title{Calculate distance to next station}
\usage{
calculateDistanceToNextStation(longitude, latitude, unit = "nm")
}
\arguments{
\item{longitude}{A numeric value indicating longitude in decimal degrees.}

\item{latitude}{A numeric value indicating latitude in decimal degrees.}

\item{unit}{A character string indicating the output unit for distance. Options include \code{nm} and \code{km}, \code{nm} is the default.}
}
\description{
Calculate the distance to next station.
}
