% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculateTransitTime}
\alias{calculateTransitTime}
\title{Calculate transit time to next station.}
\usage{
calculateTransitTime(
  distanceToNextStation,
  distanceUnit = "nm",
  transitSpeed,
  speedUnit = "kts"
)
}
\arguments{
\item{distanceToNextStation}{a numeric vector of distances to next station in either
nautical miles (\code{nm}) or kilometers (\code{km}), as denoted by \code{distanceUnit},}

\item{distanceUnit}{a character vector indicating the unit of \code{distanceToNextStation}.}

\item{transitSpeed}{a numeric vector of transit speed of the vessel in either knots (\code{kts})
or kilometers per hour (\code{kph}), as indicated by \code{speedUnit}.}

\item{speedUnit}{a character vector indicating the unit of \code{transitSpeed}.}
}
\description{
Calculate the transit time to next station.
}
\author{
Chantelle Layton
}
