% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{halifaxLineMission}
\alias{halifaxLineMission}
\title{Halifax Line example mission}
\format{
a \code{data.frame} that was read in using \code{cruisePlanning::readCruiseInput()} with columns : \code{lon_dd}, \code{lat_dd}, \code{station}, \code{operation},
\code{optime}, \code{xoptime},\code{type}, \code{transit_speed}.
}
\usage{
halifaxLineMission
}
\description{
An example mission, which leaves the Bedford Institute of Oceanography, completes the Atlantic Zone Monitoring Program Halifax Line,
and then returns to Bedford Institute of Oceaongraphy.
}
\keyword{datasets}
