% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeCruisePlanningMap}
\alias{makeCruisePlanningMap}
\title{Create cruise map}
\usage{
makeCruisePlanningMap(
  data,
  markerColumn = NULL,
  labelColumn = NULL,
  markerLabelFactor = NULL,
  NAFOShpFile,
  MPAShpFile,
  ClosureShpFile
)
}
\arguments{
\item{data}{a data.frame containing cruise planning variables, either from running \code{runCruisePlanning}, or reading
in the data file using \code{readCruiseInput}.}

\item{markerColumn}{a character vector of length one indicating which column in \code{data} to color code the markers. Default is \code{NULL},
meaning all coordinates will be the same colour.}

\item{labelColumn}{a character vector of length one indicating which column in \code{data} to use as labels on the map
for the given indicated factor in \code{markerLabelFactor} for the \code{markerColumn}.
Default is \code{NULL}, meaning label each station with a sequential number. If \code{FALSE}, no labels will be placed at each station.}

\item{markerLabelFactor}{a character vector indicating which factors in \code{markerColumn} to label on the map.
Default is \code{NULL}, meaning put labels for all factors.}

\item{NAFOShpFile}{a connection or a character string giving the name of the \code{.shp} file associated with the
NAFO zones. See details on where to download this file.}

\item{MPAShpFile}{a connection or a character string giving the name of the \code{.shp} file associated
with marine protected areas. See details on where to download this file.}

\item{ClosureShpFile}{a connection or a character string giving the name of the \code{.shp} file associated
with Other Effective Area-Based Conservation Measures (OECM). See details on where to download this file.}
}
\description{
Create cruise map
}
\details{
\itemize{
\item NAFO \code{.shp} file can be downloaded from \url{https://www.nafo.int/Data/GIS} (accessed on 2024-05-15), click on \code{Divisions}.
\item MPA \code{.shp} file can be downloaded from \url{https://open.canada.ca/data/en/dataset/a1e18963-25dd-4219-a33f-1a38c4971250}
(accessed on 2024-05-15). Under \verb{Data and Resources}, navigate to the \verb{Oceans Act Marine Protected Areas} with \code{dataset} and \code{SHP}
listed as keywords. Click on \code{Explore}, then click \verb{Go to resource} and it will download.
\item Closure \code{.shp} file can be downloaded from \url{https://open.canada.ca/data/en/dataset/44769543-7a23-4991-a53f-c2cf7c7a946f}
(accessed on 2024-05-15). Under \verb{Data and Resources}, navigate to the \verb{Other Effective Area-Based Conservation Measures} with \code{dataset} and \code{SHP}
listed as keywords. Click on \code{Explore}, then click \verb{Go to resource} and it will download.
}
}
\author{
Chantelle Layton
}
