% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readCruiseInput}
\alias{readCruiseInput}
\title{Read cruise input file}
\usage{
readCruiseInput(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to load.}
}
\value{
a data.frame
}
\description{
Read in mission input file required for \code{runCruisePlanning}. It will
check for required columns. See details for required columns.
}
\section{file}{

At a minimum, the supplied \code{.csv} file should include the following variables
\tabular{ll}{
\strong{variable name} \tab \strong{variable description} \cr
lon_dd \tab A numeric value indicating longitude in decimal degrees. \cr
lat_dd \tab A numeric value indicating latitude in decimal degrees. \cr
optime \tab A numeric value indicating the total expected operational time in hours at given lon_dd, lat_dd. \cr
transit_speed \tab A numeric value indicating the transit speed of the vessel.
}
Additional variables that are suggested (but not limited to):
\tabular{ll}{
\strong{variable name} \tab \strong{variable description} \cr
xoptime \tab A numeric value indicating any extra operational time in hours at given lon_dd, lat_dd. \cr
station \tab A character string indicating the name of the station. \cr
operation \tab An character string indicating the specific operation order for the given operation type (e.g. Net_CTD, Net_CTD_Argo, Net_2_CTD). \cr
type \tab A character string indicating the operation type (e.g. \code{operation}, \code{transit}).\cr
}
}

\author{
Chantelle Layton
}
