% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runCruisePlanning}
\alias{runCruisePlanning}
\title{Run cruise planning.}
\usage{
runCruisePlanning(
  file,
  startTime,
  writeCsv = TRUE,
  writeSubsetCsv = TRUE,
  subsetNames = c("lon_dd", "lat_dd", "operation", "station", "depth", "dist_to_next",
    "arrival_time", "departure_time")
)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to load.}

\item{startTime}{a \code{POSIXct} format of the start date time. Suggested to use \code{ISOdatetime()} for simplicity.}

\item{writeCsv}{a logical value, if \code{TRUE} a \code{.csv} file will be written in the working directory using the provided
filename, with the current local date and time appended to the end of the filename.}

\item{writeSubsetCsv}{a logical value, if \code{TRUE} a \code{.csv} file will be written in the working directory using the provided
filename, with the current local date and time appended to the end of the filename using provided names in \code{abbreviatedNames}}

\item{subsetNames}{a vector of names in provided filename to be output into a \code{.csv}. Ignored if \code{writeAbbreviatedCsv = FALSE}.
Defaults are \code{c('lon_dd', 'lat_dd','station','operation','depth','dist_to_next', 'arrival_time','departure_time')}. The
output file will provide variable names in order provided. During a cruise, this is what is readily printed off and circulated to staff.}
}
\value{
a data.frame with added calculated variables, which includes \code{dist_to_next}, \code{transit_time}, \code{arrival_time}, and \code{departure_time}
}
\description{
Run cruise planning. The distance to next set of coordinate, \code{dist_to_next}, transit time to next set of coordinates,
\code{transit_time}, arrival time on the coordinates, \code{arrival_time}, departure time at the coordinates, \code{departure_time},
and the depth at the set of coordinates is added to the provided
}
\author{
Chantelle Layton
}
