% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeCruisePlanningMap}
\alias{makeCruisePlanningMap}
\title{Create cruise map}
\usage{
makeCruisePlanningMap(
  data,
  markerColumn = NULL,
  labelColumn = NULL,
  markerLabelFactor = NULL,
  markerPopupNames = NULL,
  NAFOShpFile = NULL,
  MPAShpFile = NULL,
  ClosureShpFile = NULL,
  saveMap = FALSE,
  mapFilename = NULL
)
}
\arguments{
\item{data}{a data.frame containing cruise planning variables, either from running \code{runCruisePlanning}, or reading
in the data file using \code{readCruiseInput}.}

\item{markerColumn}{a character vector of length one indicating which column in \code{data} to color code the markers. Default is \code{NULL},
meaning all coordinates will be the same colour.}

\item{labelColumn}{a character vector of length one indicating which column in \code{data} to use as labels on the map
for the given indicated factor in \code{markerLabelFactor} for the \code{markerColumn}.}

\item{markerLabelFactor}{a character vector indicating which factors in \code{markerColumn} to label on the map.
Default is \code{NULL}, meaning put labels for all factors.}

\item{markerPopupNames}{a character vector of desired columns in \code{data} to display in popup for each station.
Default is \code{NULL}, meaning include information from all columns in \code{data}. If \code{FALSE}, no popup will show.
Default is \code{NULL}, meaning label each station with a sequential number. If \code{FALSE}, no labels will be placed at each station.}

\item{NAFOShpFile}{optional, a connection or a character string giving the name of the \code{.shp} file associated with the
NAFO zones. See details on where to download this file. Default is \code{NULL}.}

\item{MPAShpFile}{optional, a connection or a character string giving the name of the \code{.shp} file associated
with marine protected areas. See details on where to download this file. Default is \code{NULL}.}

\item{ClosureShpFile}{optional, a connection or a character string giving the name of the \code{.shp} file associated
with Other Effective Area-Based Conservation Measures (OECM). See details on where to download this file. Default is \code{NULL}.}

\item{saveMap}{a logical value indicating if the \code{leaflet} map should be saved, \code{TRUE}, or not, \code{FALSE}. Default is \code{FALSE}.}

\item{mapFilename}{a character vector of length one of the file name for the map to be saved. If left \code{NULL},
and \code{saveMap = TRUE}, a filename will be constructed following the convention \verb{leaflet_format(Sys.time(), '\%Y\%m\%d_\%H\%M').html}.}
}
\description{
Create cruise map
}
\details{
\itemize{
\item NAFO \code{.shp} file can be downloaded from \url{https://www.nafo.int/Data/GIS} (accessed on 2024-05-15), click on \code{Divisions}.
\item MPA \code{.shp} file can be downloaded from \url{https://open.canada.ca/data/en/dataset/a1e18963-25dd-4219-a33f-1a38c4971250}
(accessed on 2024-05-15). Under \verb{Data and Resources}, navigate to the \verb{Oceans Act Marine Protected Areas} with \code{dataset} and \code{SHP}
listed as keywords. Click on \code{Explore}, then click \verb{Go to resource} and it will download.
\item Closure \code{.shp} file can be downloaded from \url{https://open.canada.ca/data/en/dataset/44769543-7a23-4991-a53f-c2cf7c7a946f}
(accessed on 2024-05-15). Under \verb{Data and Resources}, navigate to the \verb{Other Effective Area-Based Conservation Measures} with \code{dataset} and \code{SHP}
listed as keywords. Click on \code{Explore}, then click \verb{Go to resource} and it will download.
}
}
\author{
Chantelle Layton
}
