% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readCruiseInput}
\alias{readCruiseInput}
\title{Read cruise input file}
\usage{
readCruiseInput(file)
}
\arguments{
\item{file}{a connection or a character string giving the name of the file to load}
}
\value{
a data.frame
}
\description{
Read in mission input file required for \code{runCruisePlanning}. It will
check for required columns.
}
\author{
Chantelle Layton
}
