% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{runCruisePlanning}
\alias{runCruisePlanning}
\title{Run cruise planning.}
\usage{
runCruisePlanning(
  file,
  startTime,
  inferDepth = TRUE,
  lonrange,
  latrange,
  writeCsv = TRUE,
  writeSubsetCsv = TRUE,
  subsetNames = c("lon_dd", "lat_dd", "operation", "station", "depth", "dist_to_next",
    "arrival_time", "departure_time")
)
}
\arguments{
\item{file}{a connection or a character string giving the name of the \code{.csv} file to load or a \code{data.frame} that contains the minimum
set of variables. See details on the minimum and suggested variables.}

\item{startTime}{a \code{POSIXct} format of the start date time. Suggested to use \code{ISOdatetime()} for simplicity.}

\item{inferDepth}{a logical value, if \code{TRUE} topographic data (ETOPO1) are downloaded and cached in the working directory and is used to infer
the depth at each set of coordinates.}

\item{lonrange}{an optional numeric vector of length 2 indicating the range of longitude (west, east) to use to
download topographic data if \code{inferDepth = TRUE}, otherwise ignored. This is suggested if continuously running scenarios
during a mission while at sea.}

\item{latrange}{an optional numeric vector of length 2 indicating the range of latitude (south, north) to use to
download topographic data if \code{inferDepth = TRUE}, otherwise ignored. This is suggested if continuously running scenarios
during a mission while at sea.}

\item{writeCsv}{a logical value, if \code{TRUE} a \code{.csv} file will be written in the working directory using the provided
filename, with the current local date and time appended to the end of the filename.}

\item{writeSubsetCsv}{a logical value, if \code{TRUE} a \code{.csv} file will be written in the working directory using the provided
filename, with the current local date and time appended to the end of the filename using provided names in \code{subsetNames}.The
output file will provide variable names in order provided. During a cruise, this is what is readily printed off and circulated to staff.}

\item{subsetNames}{a vector of names in provided \code{file} to be output into a \code{.csv}. Ignored if \code{writeSubsetCsv = FALSE}.
Defaults are \code{c('lon_dd', 'lat_dd','station','operation','depth','dist_to_next', 'arrival_time','departure_time')}.}
}
\value{
a data.frame with added calculated variables, which includes \code{dist_to_next}, \code{transit_time}, \code{arrival_time}, and \code{departure_time}
}
\description{
Run cruise planning. The distance to next set of coordinate, \code{dist_to_next}, transit time to next set of coordinates,
\code{transit_time}, arrival time on the coordinates, \code{arrival_time}, departure time at the coordinates, \code{departure_time},
and the depth at the set of coordinates is added to the provided variables in the supplied \code{file}.
}
\details{
The following provides additional details for what to include in the provided \code{.csv} file.
}
\section{file}{

At a minimum, the supplied \code{.csv} file should include the following variables
\tabular{ll}{
\strong{name} \tab \strong{description} \cr
lon_dd \tab A numeric value indicating longitude in decimal degrees. \cr
lat_dd \tab A numeric value indicating latitude in decimal degrees. \cr
optime \tab A numeric value indicating the total expected operational time in hours at given lon_dd, lat_dd. \cr
transit_speed \tab A numeric value indicating the transit speed of the vessel.
}
Additional variables that are suggested (but not limited to):
\tabular{ll}{
\strong{name} \tab \strong{description} \cr
xoptime \tab A numeric value indicating any extra operational time in hours at given lon_dd, lat_dd. \cr
station \tab A character string indicating the name of the station. \cr
operation \tab An character string indicating the specific operation order for the given operation type (e.g. Net_CTD, Net_CTD_Argo, Net_2_CTD). \cr
type \tab A character string indicating the operation type (e.g. \code{operation}, \code{transit}).\cr
}
}

\examples{
library(cruisePlanning)
data(halifaxLineMission)
startTime <- ISOdatetime(year = 2024,
                         month = 4,
                         day = 18,
                         hour = 6,
                         min = 0,
                         sec = 0,
                         tz = "" # leaving it as "" denotes the current time zone, can set to 'UTC'
                         )
d <- runCruisePlanning(file = halifaxLineMission,
                       startTime = startTime,
                       inferDepth = FALSE, # for example purposes
                       writeCsv = FALSE, # for example purposes
                       writeSubsetCsv = FALSE # for example purposes
                       )


}
\author{
Chantelle Layton
}
